- Added ORMMemoryDatabase, it imports and loads the memory tables, can fetch an entity by id and fetch table columns; - SubObjects can now be fetched in unit tests; - Added IsDirty support for MemoryDatabase; - SQLBuilder SqlParameters now store the SourceColumn; - Added (single) primary key support for unit tests from the MemoryDatabase; - Added some new unit tests, although not all work yet untill the refactor is complete and all bugs are fixed; - Replaced some ORMUtilities methods to then new SQLHelper class; - Other small changes and clean-ups.
