+++ ORMSolution/ORM/ORM/ORMPrimaryKeyIdentification.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Data;
 using System.Linq;
 
-namespace ORM.ORM
+namespace ORM
 {
     internal struct ORMPrimaryKeyIdentification
     {
