+++ ORMSolution/ORMNUnit/MemoryTables/ORGANISATIONS.xml
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
-<data>
+<DATA>
   <ORGANISATIONS Id="1" Name="The Boring Company" />
   <ORGANISATIONS Id="2" Name="SpaceX" />
-</data>
+</DATA>
